Refactored checkout.js into checkout_functions.js and checkout.js
